Release/1.0.0

+++ src/component/CategoryList/component/component/SubCategory.jsx
@@ -1,6 +1,6 @@
-import React from 'react';
-import './style.css';
-import { Link } from 'react-router-dom';
+import React, { useCallback } from 'react';
+import { useHistory } from 'react-router-dom';
+import style from './SubCategory.css';
 
 const SubCategory = ({
   name,
@@ -8,41 +8,69 @@ const SubCategory = ({
   id,
   status,
   deleteSubCategory,
+  deleteCategoryWithTodo,
+  editNameSubCategory,
   nameCategory,
 }) => {
+  const history = useHistory();
+
   const setStatusSubCategory = (e) => {
     showTodo(e.target.dataset.name, Number(e.target.dataset.idsubcategory));
+    history.push(`/category/${id}`);
   };
+  const editSubCategory = useCallback(
+    (e) => {
+      editNameSubCategory(
+        prompt('change name subCategory', e.target.dataset.subname) ||
+          e.target.dataset.subname,
+        Number(e.target.dataset.id),
+      );
+    },
+    [editNameSubCategory],
+  );
+
   const delSubCategory = (e) => {
-    deleteSubCategory(
-      Number(e.target.dataset.id),
-      e.target.dataset.namecategory,
-    );
+    // eslint-disable-next-line no-restricted-globals,no-unused-expressions
+    confirm('Are you sure to do this?') &&
+      deleteSubCategory(
+        Number(e.target.dataset.id),
+        e.target.dataset.namecategory,
+      ) &&
+      deleteCategoryWithTodo(Number(e.target.dataset.id));
   };
   return (
     <li>
-      <div className="row">
-        <div className="col-6 edit">
-          <Link className="linkRoute" to={`/category/subcategory/${id}`}>
+      <div className={style.container}>
+        <div className="edit">
+          <div className={style.linkRoute}>
             <span
               role="link"
               aria-hidden="true"
               data-name={name}
               data-idsubcategory={id}
               className={
-                status ? 'nameSubCategory colorSubCategory' : 'nameSubCategory'
+                status
+                  ? `${style.nameSubCategory} ${style.colorSubCategory}`
+                  : `${style.nameSubCategory}`
               }
               onClick={setStatusSubCategory}
             >
               {name}
             </span>
-          </Link>
-          <button type="button" data-id={id}>
+          </div>
+        </div>
+        <div className="trash">
+          <button
+            className={style.edit}
+            type="button"
+            data-subname={name}
+            data-id={id}
+            onClick={editSubCategory}
+          >
             edit
           </button>
-        </div>
-        <div className="col-6 trash">
           <button
+            className={style.del}
             type="button"
             data-id={id}
             data-namecategory={nameCategory}
